Added cvector_reserve macro
Request a change in capacity
Requests that the vector capacity be at least enough
to contain n elements.

If n is greater than the current vector capacity, the function
causes the container to reallocate its storage increasing its
capacity to n (or greater).

In all other cases, the function call does not cause a reallocation
and the vector capacity is not affected.

This function has no effect on the vector size and cannot alter its
elements.