Add copyright and license on the top of the cvector.h file (#42)
For some people who want to use this library on the other project
by just copying the file, need to add copyright on the top of the file.
If the copyright and the license are already specified on the top of the file,
this library can be much easier to be used.