Revert cvector_at changes (#76)
It is more idiomatic for cvector_at to return a reference. This allows
mutability of the element at the given index, and allows for the
existence of an element at a given index to be expressed in a different
way than explicitly checking the size of the vector.

This also corrects the documentation for cvector_front and cvector_back,
as those were still documented as returning references, not values.