Replace CVECTOR_LOGARITHMIC_GROWTH with CVECTOR_LINEAR_GROWTH (#77)
In my own usage and in the examples I can find with using this library,
it is more common to want to use the logarithmic growth algorithm than
the linear growth algorithm. This is also the default (or only option)
with vector and dynamic array data structures that are available in
other programming languages. Setting the logarithmic algorithm to the
default will reduce the burden on library users to define a special
value in each compilation unit they use the library, and reduce the risk
that they will forget to define it and lose performance.

This change is backwards compatible with any existing programs that
defined the CVECTOR_LOGARITHMIC_GROWTH value; their program will
continue to operate as intended, and they can remove the definition
if/when they want.

This change /will/ modify the behavior of programs that did not define
the CVECTOR_LOGARITHMIC_GROWTH value; they must update their program to
define the new CVECTOR_LINEAR_GROWTH value instead.